The two necessary sensors for detection, driven by python code
The two files attached are code to read values from a rotary encoder to determine the circumference of the band, and a temperature and humidity sensor to determine cold spots (possible blood clots) on the user's leg. These files connect to my firebase database in the google cloud and write to the values there, in real time.